dev-requirements.txt
pytest-cov==2.5.1
requests_mock==1.4.0
httmock==1.2.3
lockfile==0.10.2
coverage==3.7.1
mock==1.0.1
pytest-flask==0.8.1
moto==0.4.5
sphinxcontrib-httpdomain==1.3.0

# force install of sqlalchemy 0.9.9 because indexd installs 1.0.8
# this new version is not backwards compatible with 0.9.9
SQLAlchemy==0.9.9
codacy-coverage
Sphinx==1.6.5
sphinx_rtd_theme
-e git+https://github.com/uc-cdis/doiclient.git@1.0.0#egg=doiclient
-e git+https://github.com/uc-cdis/dosclient.git@1.0.0#egg=dosclient
-e git+https://git@github.com/uc-cdis/cdisutils-test.git@0.2.2#egg=cdisutilstest
-e git+https://git@github.com/uc-cdis/indexd.git@010882355a773824dfc9ba2407c8f34327e3a2a0#egg=indexd
# these two are required by fence -> cirrus -> google_cloud
pyasn1-modules==0.0.11
urllib3==1.22

-e git+https://github.com/uc-cdis/cirrus.git@0.0.0#egg=cirrus-0.0.0
-e git+https://git@github.com/uc-cdis/fence.git@1.1.1#egg=fence
